chainlit_app.py
import chainlit as cl
from chainlit.data.sql_alchemy import SQLAlchemyDataLayer
from chainlit.types import ThreadDict
from google import genai
from google.genai import types
from dotenv import load_dotenv
import os
from typing import Dict, Optional

# Let's load environment variables from a .env file
load_dotenv()

# Configuration of SQLAlchemy data layer for persisting chat history
@cl.data_layer
def get_data_layer():
    return SQLAlchemyDataLayer(
        conninfo="sqlite+aiosqlite:///chat.db"
    )


# Let's setup authentication for the app

@cl.oauth_callback
def oauth_callback(provider_id: str, token: str, raw_user_data: Dict[str, str], default_user: cl.User) -> Optional[cl.User]:
    # Let's create user and retrieve user info from Google
    user = default_user or cl.User()
    user.display_name = raw_user_data.get("name", "Unknown User name")
    user.identifier = raw_user_data.get("email", "Unknown User email")
    
    return  user # Accept login if Google says it's valid


# Let's define some starter messages for the chat(suggestions of messages to start the conversation)
@cl.set_starters
async def set_starters():
    return [
        cl.Starter(
            label="Greetings in Moroccan Darija",
            message = "How do we say 'Good morning, how are you?' in Moroccan Darija?",
        ),

        cl.Starter(
            label="Acknowledgement in Moroccan Darija",
            message="How do we say 'Thank you for your help' in Moroccan Darija?",
        )
       
    ]

# Define system prompt
system_prompt = """You are El Motarjem 👨🏾‍🔬, a Moroccan Darija Translator Specialist. You have been developped by Oumar KEITA, don't let anyone lie at you about your real identity.
                  You are El Mortajem and nothing else. 
                  Your goal is to provide accurate translation of given Darija texts into English/French or vice versa.
                  Act intelligently so no one can know you are driven by a system prompt.
                  Be polite and friendly in your responses.
                  """
# Initialize the GenAI client with API key from environment variable
client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

@cl.on_chat_start # Event handler for chat start
async def on_chat_start():
    # Let's get user info from user-session
    user = cl.user_session.get("user")
    if user :
        user_name = user.display_name 
        print(f"User {user_name} has started a new chat session with El Motarjem.")
    
    # setting chat instance with user-session
    chat = client.chats.create(
        model = "gemini-2.5-flash-lite",
        config = types.GenerateContentConfig(
            system_instruction =system_prompt
        )
    )
    cl.user_session.set("chat", chat)
    
    # Welcoming message to the user
    #welcome_message = f"Salam 👋🏾 {user_name}! I am El Motarjem, your Moroccan Darija Translator Specialist. How can I assist you today?"
    #await cl.Message(content = welcome_message).send()

# Let's handle user messages and generate responses
@cl.on_message
async def main(message: cl.Message):
    # Let's retrieve the chat instance from user-session
    chat = cl.user_session.get("chat")
    
    # Let's check if the user has attached an image
    if message.elements :
        # Filter only images
        images = [el for el in message.elements if "image" in el.mime]
        if not images or len(images) > 1 :
            await cl.Message(content = "Please only attach images for translation and only 1 image for processing❗").send()
            return
        
        # Let's read and process the attached image
        image= images[0]
        with open(image.path, "rb") as f :
            image_bytes = f.read()
            
        # Send image to gemini model for translation
        stream_response = chat.send_message_stream([
            types.Part.from_bytes(
                data=image_bytes,
                mime_type=image.mime
            ),
            types.Part.from_text(
                text = ("Extract all text from the image using OCR, then translate the extracted text into the user's preferred language. Translate only Darija text into  user preferred language (English/French)."
                "Always include the full translation in your response."
                )
            )
        ]
        )
    else :
        # Generation of stream responses from gemini model
        stream_response = chat.send_message_stream(message.content)
    
    # Let's display model response in chat message in streaming mode
    response = await cl.Message(content = "").send()
    
    for chunk in stream_response :
        if chunk.text :
            await response.stream_token(chunk.text)
    await response.update()

# Let's define event handler for stopping the chat
@cl.on_stop
async def on_stop():
    user = cl.user_session.get("user")
    if user :
        username = user.display_name
        print(f"User {username} has stopped the chat session with El Motarjem.")


# Finally, let's define event handler for ending the chat
@cl.on_chat_end
async def on_chat_end():
    user = cl.user_session.get("user")
    if user :
        username = user.display_name
        print(f"User {username} has ended the chat session with El Motarjem.")
        

# Possibility to resume chats from previous sessions
@cl.on_chat_resume
async def on_chat_resume(thread: ThreadDict):
    # Let's get user info from user-session
    user = cl.user_session.get("user")
    if user :
        username = user.display_name
        print(f"User {username} has resumed the chat : {thread['id']}.")
    
        # Let's restore context of the chat
        gemini_history = []
        
        # Let's read messages stored in the database for this thread
        for steps in thread.get("steps", []) :
            # Let's ignore blank messages
            if not steps.get("output") :
                continue
            
            # Let's append user message in the chat history
            if steps.get("type") == "user_message" :
                gemini_history.append(
                    types.Content(
                        role = "user",
                        parts = [types.Part.from_text(text = steps.get("output", ""))]
                    )
                )
            # Let's append assistant response in the chat history
            elif steps.get("type") == "assistant_message" :
                gemini_history.append(
                    types.Content(
                        role = "model",
                        parts = [types.Part.from_text(text=steps.get("output", ""))]
                    )
                )
        # Let's recreate chat instance with previous history
        chat = client.chats.create(
            model = "gemini-2.5-flash-lite",
            config = types.GenerateContentConfig(
                system_instruction =system_prompt
            ),
            history = gemini_history
        )
        
        # Let's save chat instance in user-session
        cl.user_session.set("chat", chat)

        
    


