requirements.txt
chainlit
google-genai
python-dotenv
aiosqlite
sqlalchemy
